fix(solver): split indexed access inference helper
