+++ crates/tsz-solver/src/operations/constraints/mod.rs
@@ -5,6 +5,7 @@
 //! recursive traversal of complex type structures (objects, functions, tuples,
 //! conditionals, mapped types, etc.) to extract inference candidates.
 
+mod indexed_access;
 mod reverse_mapped;
 mod signatures;
 mod walker;
